ChripCSVDBService handles the limit of cheeps read
The functionallity of the limit was before in Chirp.CLI.Client
But is now in ChirpCSVDBService. The Read methode is responsable for
limiting the amount of cheeps returned fro the http get call.
The http call is now sending success or fail responses back, which is
then displayed to the user.